+++ packages/cms-cli/lib/createFunctionPrompt.js
